+++ envolved/__init__.py
@@ -1,6 +1,6 @@
 from envolved._version import __version__
 from envolved.describe import describe_env_vars
-from envolved.envvar import EnvVar, as_default, discard, env_var, inferred_env_var, missing, no_patch
+from envolved.envvar import EnvVar, Factory, as_default, discard, env_var, inferred_env_var, missing, no_patch
 from envolved.exceptions import MissingEnvError
 
 __all__ = [
@@ -14,4 +14,5 @@ __all__ = [
     "inferred_env_var",
     "missing",
     "no_patch",
+    "Factory",
 ]